[refactor] dto 분리, 가드 적용
