+++ libs/shared-types/src/index.ts
@@ -48,6 +48,7 @@ export * from './lib/dto/findBroadcaster.dto';
 export * from './lib/dto/findExports.dto';
 export * from './lib/dto/findFmOrderDetails.dto';
 export * from './lib/dto/FindFmOrders.dto';
+export * from './lib/dto/findNotifications.dto';
 export * from './lib/dto/findSeller.dto';
 export * from './lib/dto/findSettlementHistory.dto';
 export * from './lib/dto/goodsConfirmation.dto';
@@ -59,6 +60,7 @@ export * from './lib/dto/goodsOptionSupply.dto';
 export * from './lib/dto/inquiry.dto';
 export * from './lib/dto/liveShopping.dto';
 export * from './lib/dto/loginUser.dto';
+export * from './lib/dto/markNotificationReadState.dto';
 export * from './lib/dto/passwordValidate.dto';
 export * from './lib/dto/post.dto';
 export * from './lib/dto/registGoods.dto';
